# apps/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/autenticacao/mobile/pareamentos-qr/resgatar": {
      "post": {
        "operationId": "resgatarPareamentoQrMobile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaResgatePareamentoQrDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResgatePareamentoQrDto"
                }
              }
            }
          }
        },
        "summary": "Resgata uma vez o token efêmero de pareamento QR",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/pareamentos-qr/consultar": {
      "post": {
        "operationId": "consultarPareamentoQrMobile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaComprovantePareamentoQrDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoPareamentoQrMobileDto"
                }
              }
            }
          }
        },
        "summary": "Consulta a confirmação web do pareamento resgatado",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/pareamentos-qr/concluir": {
      "post": {
        "operationId": "concluirPareamentoQrMobile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaComprovantePareamentoQrDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoMobileDto"
                }
              }
            }
          }
        },
        "summary": "Conclui o pareamento confirmado e entrega sessão somente ao mobile",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/entrar": {
      "post": {
        "operationId": "entrarSessaoMobile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaLoginMobileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoMobileDto"
                }
              }
            }
          }
        },
        "summary": "Autentica e vincula uma sessão mobile ao aparelho",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/renovar": {
      "post": {
        "operationId": "renovarSessaoMobile",
        "parameters": [
          {
            "name": "x-segredo-dispositivo",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dispositivo-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntradaRefreshMobileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoMobileDto"
                }
              }
            }
          }
        },
        "summary": "Rotaciona access e refresh tokens mobile",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/dispositivos": {
      "get": {
        "operationId": "listarDispositivosMobile",
        "parameters": [
          {
            "name": "x-segredo-dispositivo",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dispositivo-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResumoDispositivoMobileDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoMobile": []
          }
        ],
        "summary": "Lista os dispositivos móveis ativos do usuário atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/dispositivos/{dispositivoId}/revogar": {
      "post": {
        "operationId": "revogarDispositivoMobileDoUsuario",
        "parameters": [
          {
            "name": "dispositivoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-segredo-dispositivo",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dispositivo-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoMobile": []
          }
        ],
        "summary": "Revoga um dispositivo móvel do usuário atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/sessao": {
      "get": {
        "operationId": "obterSessaoMobile",
        "parameters": [
          {
            "name": "x-segredo-dispositivo",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dispositivo-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextoSessaoMobileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessaoMobile": []
          }
        ],
        "summary": "Valida a sessão mobile atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
    "/api/v1/autenticacao/mobile/sair": {
      "post": {
        "operationId": "sairSessaoMobile",
        "parameters": [
          {
            "name": "x-segredo-dispositivo",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-dispositivo-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "sessaoMobile": []
          }
        ],
        "summary": "Revoga a sessão mobile atual",
        "tags": [
          "autenticacao"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Vyntra Omnichannel",
    "description": "Contrato público e versionado do Omnichannel V1.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "sessaoWeb": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-vyntra_sessao"
      },
      "sessaoMobile": {
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "type": "http"
      }
    },
    "schemas": {
...
      "EntradaResgatePareamentoQrDto": {
        "type": "object",
        "properties": {
          "token_qr": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "identificador_instalacao": {
            "type": "string",
            "format": "uuid"
          },
          "segredo_vinculo": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "plataforma": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "modelo_sanitizado": {
            "type": "string",
            "maxLength": 120
          },
          "versao_aplicativo": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "token_qr",
          "identificador_instalacao",
          "segredo_vinculo",
          "plataforma",
          "versao_aplicativo"
        ]
      },
      "ResgatePareamentoQrDto": {
        "type": "object",
        "properties": {
          "pareamento_id": {
            "type": "string",
            "format": "uuid"
          },
          "comprovante_resgate": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "pareamento_id",
          "comprovante_resgate",
          "expira_em"
        ]
      },
      "EntradaComprovantePareamentoQrDto": {
        "type": "object",
        "properties": {
          "pareamento_id": {
            "type": "string",
            "format": "uuid"
          },
          "comprovante_resgate": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "identificador_instalacao": {
            "type": "string",
            "format": "uuid"
          },
          "segredo_vinculo": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "plataforma": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "modelo_sanitizado": {
            "type": "string",
            "maxLength": 120
          },
          "versao_aplicativo": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "pareamento_id",
          "comprovante_resgate",
          "identificador_instalacao",
          "segredo_vinculo",
          "plataforma",
          "versao_aplicativo"
        ]
      },
      "EstadoPareamentoQrMobileDto": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "enum": [
              "AGUARDANDO_CONFIRMACAO",
              "CONFIRMADO"
            ]
          },
          "expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "estado",
          "expira_em"
        ]
      },
      "SessaoMobileDto": {
        "type": "object",
        "properties": {
          "sessao_id": {
            "type": "string",
            "format": "uuid"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid"
          },
          "dispositivo_id": {
            "type": "string",
            "format": "uuid"
          },
          "nome_exibicao": {
            "type": "string"
          },
          "token_acesso": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "token_refresh": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "acesso_expira_em": {
            "type": "string",
            "format": "date-time"
          },
          "refresh_expira_em": {
            "type": "string",
            "format": "date-time"
          },
          "dispositivo_substituido": {
            "type": "boolean"
          }
        },
        "required": [
          "sessao_id",
          "usuario_id",
          "dispositivo_id",
          "nome_exibicao",
          "token_acesso",
          "token_refresh",
          "acesso_expira_em",
          "refresh_expira_em",
          "dispositivo_substituido"
        ]
      },
      "EntradaLoginMobileDto": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "string",
            "example": "maria.silva",
            "maxLength": 120,
            "minLength": 3
          },
          "senha": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12
          },
          "identificador_instalacao": {
            "type": "string",
            "format": "uuid"
          },
          "segredo_vinculo": {
            "type": "string",
            "minLength": 43,
            "maxLength": 43
          },
          "plataforma": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "modelo_sanitizado": {
            "type": "string",
            "maxLength": 120
          },
          "versao_aplicativo": {
            "type": "string",
            "example": "1.0.0",
            "maxLength": 40
          }
        },
        "required": [
          "identificador",
          "senha",
          "identificador_instalacao",
          "segredo_vinculo",
          "plataforma",
          "versao_aplicativo"
        ]
      },
      "EntradaRefreshMobileDto": {
        "type": "object",
        "properties": {
          "token_refresh": {
            "type": "string",
            "maxLength": 43,
            "minLength": 43
          }
        },
        "required": [
          "token_refresh"
        ]
      },
      "ResumoDispositivoMobileDto": {
        "type": "object",
        "properties": {
          "dispositivo_id": {
            "type": "string",
            "format": "uuid"
          },
          "atual": {
            "type": "boolean"
          },
          "plataforma": {
            "type": "string",
            "enum": [
              "IOS",
              "ANDROID"
            ]
          },
          "modelo_sanitizado": {
            "type": "string"
          },
          "versao_aplicativo": {
            "type": "string"
          },
          "ultimo_acesso_em": {
            "type": "string",
            "format": "date-time"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "dispositivo_id",
          "atual",
          "plataforma",
          "versao_aplicativo",
          "ultimo_acesso_em",
          "criado_em"
        ]
      },
      "ContextoSessaoMobileDto": {
        "type": "object",
        "properties": {
          "sessao_id": {
            "type": "string",
            "format": "uuid"
          },
          "usuario_id": {
            "type": "string",
            "format": "uuid"
          },
          "dispositivo_id": {
            "type": "string",
            "format": "uuid"
          },
          "nome_exibicao": {
            "type": "string"
          },
          "acesso_expira_em": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "sessao_id",
          "usuario_id",
          "dispositivo_id",
          "nome_exibicao",
          "acesso_expira_em"
        ]
      },
...
    }
  }
}